Split DI logic into own class
